core(network-analyzer): move throughput to NetworkAnalyzer
