+++ lighthouse-core/test/results/sample_v2.json
@@ -1825,53 +1825,43 @@
         "items": [
           {
             "url": "http://localhost:10200/zone.js",
-            "totalBytes": 71654,
-            "totalMs": 409.9376550703614
+            "totalBytes": 71654
           },
           {
             "url": "http://ajax.googleapis.com/ajax/libs/jquery/2.1.1/jquery.min.js",
-            "totalBytes": 30174,
-            "totalMs": 172.62761051850677
+            "totalBytes": 30174
           },
           {
             "url": "http://localhost:10200/dobetterweb/lighthouse-480x318.jpg",
-            "totalBytes": 24741,
-            "totalMs": 141.54502922510693
+            "totalBytes": 24741
           },
           {
             "url": "http://localhost:10200/dobetterweb/dbw_tester.html",
-            "totalBytes": 12640,
-            "totalMs": 72.31434337356418
+            "totalBytes": 12640
           },
           {
             "url": "http://localhost:10200/dobetterweb/dbw_tester.html",
-            "totalBytes": 12640,
-            "totalMs": 72.31434337356418
+            "totalBytes": 12640
           },
           {
             "url": "http://localhost:10200/dobetterweb/dbw_tester.js",
-            "totalBytes": 1703,
-            "totalMs": 9.74298471243511
+            "totalBytes": 1703
           },
           {
             "url": "http://localhost:10200/dobetterweb/dbw_disabled.css?delay=200&isdisabled",
-            "totalBytes": 1108,
-            "totalMs": 6.338947188125721
+            "totalBytes": 1108
           },
           {
             "url": "http://localhost:10200/dobetterweb/dbw_tester.css?delay=2200",
-            "totalBytes": 821,
-            "totalMs": 4.696999676400015
+            "totalBytes": 821
           },
           {
             "url": "http://localhost:10200/dobetterweb/dbw_tester.css?delay=3000&async=true",
-            "totalBytes": 821,
-            "totalMs": 4.696999676400015
+            "totalBytes": 821
           },
           {
             "url": "http://localhost:10200/dobetterweb/dbw_tester.css?delay=2000&async=true",
-            "totalBytes": 821,
-            "totalMs": 4.696999676400015
+            "totalBytes": 821
           }
         ]
       }
